refactor(UNT-T16736): audit changes for version 1
