+++ src/components/PhotoGallery/Styles.ts
