+++ src/components/PhotoGallery/Types.ts
