+++ src/components/PhotoGallery/components/Header.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Text, TouchableOpacity } from 'react-native';
 import Animated from 'react-native-reanimated';
-import type { PhotoModalHeaderProps, PhotosModalProps } from '../types';
+import type { PhotoModalHeaderProps, PhotosModalProps } from '../Types';
 import styles from './PhotoModalStyle';
 
 const Header = ({
