+++ src/components/PhotoGallery/components/ListItem.tsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react';
 import { TouchableOpacity } from 'react-native';
 import Animated from 'react-native-reanimated';
-import type { ListItemProps } from '../types';
+import type { ListItemProps } from '../Types';
 import AnimatedImage from './AnimatedImage';
 import styles from './ListItemStyle';
