+++ src/components/PhotoGallery/index.ts
@@ -1,3 +1,4 @@
 import PhotoGallery from './PhotoGallery';
 
 export { PhotoGallery };
+export type { ArrayData, PhotoGalleryProps, ListItemProps } from './Types';
